src/requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
h5py==3.12.1
idna==3.10
imageio==2.19.3
imageio-ffmpeg==0.5.1
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
opencv-python==4.9.0.80
psycopg2-binary==2.9.10
pyarrow==18.1.0
requests==2.32.3
SQLAlchemy==2.0.36
stack-data==0.6.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
